Code/_CGNet_BSD500/requirements.txt
numpy
torch
torchvision
Pillow
matplotlib
tqdm
scikit-image
